Cleaned up the repository and added word 2 markdown testing
