--- crewai/resume-reviewer/.env.example
@@ -1,1 +1,0 @@
-SERPER_API_KEY=""
