+++ crewai/resumereviewer/.env.example
@@ -1,0 +1,2 @@
+OPENAI_API_KEY=''
+SERPER_API_KEY=''
