+++ crewai/resumereviewer/README.md
@@ -1,0 +1,7 @@
+# resume Reviewer
+The objective of this python project is to take a job description printout as markdown and align the resume accordingly to the job description.
+To make this as a python library 
+
+
+## First Attempt
+Struggled with barely any support to run it FULLY locally. So I translated the embeddings to OPENAI while the other agents used llama3 instead
